Move parameters to standalone file
