+++ src/prefect/core/__init__.py
@@ -1,3 +1,4 @@
-from prefect.core.task import Task, Parameter
-from prefect.core.flow import Flow
+from prefect.core.task import Task
+from prefect.core.parameter import Parameter
 from prefect.core.edge import Edge
+from prefect.core.flow import Flow
