+++ src/prefect/serialization/task.py
@@ -118,7 +118,7 @@ class TaskSchema(TaskMethodsMixin, ObjectSchema):
 
 class ParameterSchema(TaskMethodsMixin, ObjectSchema):
     class Meta:
-        object_class = lambda: prefect.core.task.Parameter  # type: ignore
+        object_class = lambda: prefect.core.parameter.Parameter  # type: ignore
         exclude_fields = ["type", "outputs", "slug"]
 
     type = fields.Function(lambda task: to_qualified_name(type(task)), lambda x: x)
