+++ tests/flows/test_parameter.py
@@ -1,7 +1,8 @@
 import pytest
 
 from prefect.core.flow import Flow
-from prefect.core.task import Parameter, Task
+from prefect.core.task import Task
+from prefect.core.parameter import Parameter
 from prefect.tasks.core.function import FunctionTask
